move openapi codegen into appropriate place
