+++ reflectapi-demo/src/tests/assert.rs
@@ -134,7 +134,7 @@ macro_rules! assert_snapshot {
         insta::assert_json_snapshot!(schema);
         insta::assert_snapshot!(super::into_typescript_code::<$T>());
         insta::assert_snapshot!(super::into_rust_code::<$T>());
-        insta::assert_json_snapshot!(reflectapi::openapi::Spec::from(&schema));
+        insta::assert_json_snapshot!(reflectapi::codegen::openapi::Spec::from(&schema));
     }};
 }
 
@@ -152,6 +152,6 @@ macro_rules! assert_builder_snapshot {
         insta::assert_json_snapshot!(schema);
         insta::assert_snapshot!(typescript);
         insta::assert_snapshot!(rust);
-        insta::assert_json_snapshot!(reflectapi::openapi::Spec::from(&schema));
+        insta::assert_json_snapshot!(reflectapi::codegen::openapi::Spec::from(&schema));
     }};
 }
