+++ reflectapi-demo/src/tests/mod.rs
@@ -25,7 +25,7 @@ fn write_schema() {
 fn write_openapi_spec() {
     let (schema, _) = crate::builder().build().unwrap();
 
-    let spec = reflectapi::openapi::Spec::from(&schema);
+    let spec = reflectapi::codegen::openapi::Spec::from(&schema);
     let s = serde_json::to_string_pretty(&spec).unwrap();
 
     std::fs::write(
